+++ packages/faceted-search/CHANGELOG.md
@@ -26,9 +26,18 @@ Types of changes
 
 ## [unreleased]
 
+### Breaking change
+
+- ⤴️ @talend/ui: >= 4.27.0
+
+### Fixed
+
+- [fixed](https://github.com/Talend/ui/pull/2631): Remove "Selected values only" for checkbox facet
+
 ### Added
 
-- [feat](https://github.com/Talend/ui/pull/2636/): Use Input text for Badge text instead of Textarea
+- [feat](https://github.com/Talend/ui/pull/2636): Use Input text for Badge text instead of Textarea
+- [feat](https://github.com/Talend/ui/pull/2638): use toggle for basic/advance button
 
 ## [0.2.6]